+++ src/factories/PairFactory.sol
@@ -2,6 +2,7 @@
 pragma solidity ^0.8.15;
 
 contract PairFactory {
+
     bool public isPaused;
     address public pauser;
     address public pendingPauser;
@@ -12,8 +13,7 @@ contract PairFactory {
     address public feeManager;
     address public pendingFeeManager;
 
-    mapping(address => mapping(address => mapping(bool => address)))
-        public getPair;
+    mapping(address => mapping(address => mapping(bool => address))) public getPair;
     address[] public allPairs;
     // mapping(address => bool) public isPair; // simplified check if its a pair, given that `stable` flag might not be available in peripherals
 
@@ -21,13 +21,7 @@ contract PairFactory {
     address internal _temp1;
     bool internal _temp;
 
-    event PairCreated(
-        address indexed token0,
-        address indexed token1,
-        bool stable,
-        address pair,
-        uint256
-    );
+    event PairCreated(address indexed token0, address indexed token1, bool stable, address pair, uint);
 
     constructor() {
         pauser = msg.sender;
@@ -37,7 +31,7 @@ contract PairFactory {
         volatileFee = 2;
     }
 
-    function allPairsLength() external view returns (uint256) {
+    function allPairsLength() external view returns (uint) {
         return allPairs.length;
     }
 
@@ -57,19 +51,19 @@ contract PairFactory {
     }
 
     function setFeeManager(address _feeManager) external {
-        require(msg.sender == feeManager, "not fee manager");
+        require(msg.sender == feeManager, 'not fee manager');
         pendingFeeManager = _feeManager;
     }
 
     function acceptFeeManager() external {
-        require(msg.sender == pendingFeeManager, "not pending fee manager");
+        require(msg.sender == pendingFeeManager, 'not pending fee manager');
         feeManager = pendingFeeManager;
     }
 
     function setFee(bool _stable, uint256 _fee) external {
-        require(msg.sender == feeManager, "not fee manager");
-        require(_fee <= MAX_FEE, "fee too high");
-        require(_fee != 0, "fee must be nonzero");
+        require(msg.sender == feeManager, 'not fee manager');
+        require(_fee <= MAX_FEE, 'fee too high');
+        require(_fee != 0, 'fee must be nonzero');
         if (_stable) {
             stableFee = _fee;
         } else {
@@ -77,7 +71,7 @@ contract PairFactory {
         }
     }
 
-    function getFee(bool _stable) public view returns (uint256) {
+    function getFee(bool _stable) public view returns(uint256) {
         return _stable ? stableFee : volatileFee;
     }
 
@@ -85,15 +79,7 @@ contract PairFactory {
     //     return keccak256(type(Pair).creationCode);
     // }
 
-    function getInitializable()
-        external
-        view
-        returns (
-            address,
-            address,
-            bool
-        )
-    {
+    function getInitializable() external view returns (address, address, bool) {
         return (_temp0, _temp1, _temp);
     }
