+++ src/interfaces/IMinter.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.8.15;
 
 interface IMinter {
-    function update_period() external returns (uint256);
+    function update_period() external returns (uint);
 }
